refactor: split correction service
