+++ services/corrections/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @file services/corrections/index.ts
+ * @description Central exports for correction helpers.
+ */
+export * from './base';
+export * from './character';
+export * from './item';
+export * from './map';
+export * from './name';
+export * from './dialogue';
